requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
torch
pytest==8.3.4
httpx==0.28.1
python-multipart==0.0.18
scikit-learn==1.5.2
pypdf==5.1.0
python-docx==1.1.2
